requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
dataclasses-json==0.6.3
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
langsmith==0.0.83
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.9.0
packaging==23.2
pgvector==0.2.4
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-extensions==4.9.0
typing-inspect==0.9.0
urllib3==2.1.0
yarl==1.9.4












